Some safety with move arrays, small refactor appendchar

+++ src/solvers/coord/types_macros.h
@@ -11,5 +11,5 @@ typedef struct {
 	uint32_t moves_mask;
 	uint64_t trans_mask;
 	uint8_t axistrans[3];
-	bool (*is_admissible)(int n, uint8_t [n]);
+	bool (*is_admissible)(size_t n, uint8_t [n]);
 } coord_t;
